# Optimize generated code