Generates better code for a few commonly used instructions and instruction sequences

||before|after|
|-|-|-|
|psibase-tests|10.650s|6.601s|
|TokenSys-test|2m22.099s|1m26.572s|